[Refactor] Refactor Optimizer&OptimizerContext interface (#55117)
