+++ fe/fe-core/src/main/java/com/starrocks/sql/optimizer/MaterializedViewOptimizer.java
@@ -36,24 +36,24 @@ public class MaterializedViewOptimizer {
                                   ConnectContext connectContext,
                                   boolean inlineView) {
         // optimize the sql by rule and disable rule based materialized view rewrite
-        OptimizerConfig optimizerConfig = new OptimizerConfig(OptimizerConfig.OptimizerAlgorithm.RULE_BASED);
+        OptimizerOptions optimizerOptions = OptimizerOptions.newRuleBaseOpt();
         // Disable partition prune for mv's plan so no needs  to compensate pruned predicates anymore.
         // Only needs to compensate mv's ref-base-table's partition predicates when mv's freshness cannot be satisfied.
-        optimizerConfig.disableRule(RuleType.GP_PARTITION_PRUNE);
-        optimizerConfig.disableRule(RuleType.GP_ALL_MV_REWRITE);
+        optimizerOptions.disableRule(RuleType.GP_PARTITION_PRUNE);
+        optimizerOptions.disableRule(RuleType.GP_ALL_MV_REWRITE);
         // INTERSECT_REWRITE is used for INTERSECT related plan optimize, which can not be SPJG;
         // And INTERSECT_REWRITE should be based on PARTITION_PRUNE rule set.
         // So exclude it
-        optimizerConfig.disableRule(RuleType.GP_INTERSECT_REWRITE);
-        optimizerConfig.disableRule(RuleType.TF_REWRITE_GROUP_BY_COUNT_DISTINCT);
-        optimizerConfig.disableRule(RuleType.TF_PRUNE_EMPTY_SCAN);
-        optimizerConfig.disableRule(RuleType.TF_MV_TEXT_MATCH_REWRITE_RULE);
-        optimizerConfig.disableRule(RuleType.TF_MV_TRANSPARENT_REWRITE_RULE);
-        optimizerConfig.disableRule(RuleType.TF_ELIMINATE_AGG);
-        optimizerConfig.disableRule(RuleType.TF_PULL_UP_PREDICATE_SCAN);
+        optimizerOptions.disableRule(RuleType.GP_INTERSECT_REWRITE);
+        optimizerOptions.disableRule(RuleType.TF_REWRITE_GROUP_BY_COUNT_DISTINCT);
+        optimizerOptions.disableRule(RuleType.TF_PRUNE_EMPTY_SCAN);
+        optimizerOptions.disableRule(RuleType.TF_MV_TEXT_MATCH_REWRITE_RULE);
+        optimizerOptions.disableRule(RuleType.TF_MV_TRANSPARENT_REWRITE_RULE);
+        optimizerOptions.disableRule(RuleType.TF_ELIMINATE_AGG);
+        optimizerOptions.disableRule(RuleType.TF_PULL_UP_PREDICATE_SCAN);
         // For sync mv, no rewrite query by original sync mv rule to avoid useless rewrite.
         if (mv.getRefreshScheme().isSync()) {
-            optimizerConfig.disableRule(RuleType.TF_MATERIALIZED_VIEW);
+            optimizerOptions.disableRule(RuleType.TF_MATERIALIZED_VIEW);
         }
         ColumnRefFactory columnRefFactory = new ColumnRefFactory();
         String mvSql = mv.getViewDefineSql();
@@ -77,7 +77,8 @@ public class MaterializedViewOptimizer {
         try {
             // get optimized plan of mv's defined query
             Pair<OptExpression, LogicalPlan> plans =
-                    MvUtils.getRuleOptimizedLogicalPlan(stmt, columnRefFactory, connectContext, optimizerConfig, inlineView);
+                    MvUtils.getRuleOptimizedLogicalPlan(stmt, columnRefFactory, connectContext, optimizerOptions,
+                            inlineView);
             if (plans == null) {
                 return new MvPlanContext(false, "No query plan for it");
             }
