+++ fe/fe-core/src/main/java/com/starrocks/sql/optimizer/OptExpression.java
@@ -63,8 +63,6 @@ public class OptExpression {
 
     private UKFKConstraints constraints;
 
-    private Boolean isShortCircuit = false;
-
     // the flag if its parent has required data distribution property for this expression
     private boolean existRequiredDistribution = true;
 
@@ -79,13 +77,6 @@ public class OptExpression {
     public static OptExpression create(Operator op, OptExpression... inputs) {
         OptExpression expr = new OptExpression(op);
         expr.inputs = Lists.newArrayList(inputs);
-        return expr;
-    }
-
-    public static OptExpression createForShortCircuit(Operator op, OptExpression input, boolean isShortCircuit) {
-        OptExpression expr = new OptExpression(op);
-        expr.inputs = Lists.newArrayList(input);
-        expr.setShortCircuit(isShortCircuit);
         return expr;
     }
 
@@ -229,14 +220,6 @@ public class OptExpression {
 
     public void setCost(double cost) {
         this.cost = cost;
-    }
-
-    public Boolean getShortCircuit() {
-        return isShortCircuit;
-    }
-
-    public void setShortCircuit(Boolean shortCircuit) {
-        isShortCircuit = shortCircuit;
     }
 
     @Override
