+++ fe/fe-core/src/main/java/com/starrocks/sql/optimizer/OptimizerContext.java
@@ -14,27 +14,25 @@
 
 package com.starrocks.sql.optimizer;
 
-import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Stopwatch;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import com.starrocks.catalog.OlapTable;
 import com.starrocks.qe.ConnectContext;
 import com.starrocks.qe.SessionVariable;
-import com.starrocks.server.GlobalStateMgr;
+import com.starrocks.sql.ast.StatementBase;
 import com.starrocks.sql.common.ErrorType;
 import com.starrocks.sql.common.StarRocksPlannerException;
 import com.starrocks.sql.optimizer.base.ColumnRefFactory;
 import com.starrocks.sql.optimizer.dump.DumpInfo;
-import com.starrocks.sql.optimizer.operator.logical.LogicalOlapScanOperator;
 import com.starrocks.sql.optimizer.operator.scalar.IsNullPredicateOperator;
 import com.starrocks.sql.optimizer.rewrite.JoinPredicatePushdown;
 import com.starrocks.sql.optimizer.rule.RuleSet;
 import com.starrocks.sql.optimizer.rule.RuleType;
 import com.starrocks.sql.optimizer.task.TaskContext;
 import com.starrocks.sql.optimizer.task.TaskScheduler;
+import com.starrocks.sql.optimizer.transformer.MVTransformerContext;
 
-import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -42,118 +40,137 @@ import java.util.UUID;
 import java.util.concurrent.TimeUnit;
 
 public class OptimizerContext {
-    private final UUID queryId;
-    private final Memo memo;
-    private final RuleSet ruleSet;
-    private final GlobalStateMgr globalStateMgr;
-    private final TaskScheduler taskScheduler;
-    private final ColumnRefFactory columnRefFactory;
-    private SessionVariable sessionVariable;
-    private DumpInfo dumpInfo;
-    private Set<Long> currentSqlDbIds;
-    private CTEContext cteContext;
-    private TaskContext currentTaskContext;
-    private final OptimizerConfig optimizerConfig;
-
+    // ============================ Query ============================
+    private StatementBase statement;
+    private ConnectContext connectContext;
+    private ColumnRefFactory columnRefFactory;
     private Set<OlapTable> queryTables;
-
     private long updateTableId = -1;
 
-    private boolean isObtainedFromInternalStatistics = false;
+    private OptimizerOptions optimizerOptions;
+
+    // ============================ Optimizer ============================
+    private Memo memo;
+    private final RuleSet ruleSet;
+    private TaskScheduler taskScheduler;
+
+    private final CTEContext cteContext;
+    private TaskContext currentTaskContext;
+    private final QueryMaterializationContext queryMaterializationContext = new QueryMaterializationContext();
+
+    private MVTransformerContext mvTransformerContext;
+
+    // uniquePartitionIdGenerator for external catalog
+    private long uniquePartitionIdGenerator = 0L;
     private final Stopwatch optimizerTimer = Stopwatch.createStarted();
     private final Map<RuleType, Stopwatch> ruleWatchMap = Maps.newHashMap();
 
+    // ============================ Task Variables ============================
     // The context for join predicate pushdown rule
-    private JoinPredicatePushdown.JoinPredicatePushDownContext joinPredicatePushDownContext =
+    private final JoinPredicatePushdown.JoinPredicatePushDownContext joinPredicatePushDownContext =
             new JoinPredicatePushdown.JoinPredicatePushDownContext();
     // QueryMaterializationContext is different from MaterializationContext that it keeps the context during the query
     // lifecycle instead of per materialized view.
-    private QueryMaterializationContext queryMaterializationContext = new QueryMaterializationContext();
 
-    private boolean isShortCircuit = false;
+    private boolean isObtainedFromInternalStatistics = false;
     private boolean inMemoPhase = false;
 
     // Is not null predicate can be derived from inner join or semi join,
     // which should be kept to be used to convert outer join into inner join.
-    private List<IsNullPredicateOperator> pushdownNotNullPredicates = Lists.newArrayList();
+    private final List<IsNullPredicateOperator> pushdownNotNullPredicates = Lists.newArrayList();
 
-    // uniquePartitionIdGenerator for external catalog
-    private long uniquePartitionIdGenerator = 0L;
-
-    // collect all LogicalOlapScanOperators in the query before any optimization
-    private List<LogicalOlapScanOperator> allLogicalOlapScanOperators;
-
-    @VisibleForTesting
-    public OptimizerContext(Memo memo, ColumnRefFactory columnRefFactory) {
-        this.memo = memo;
+    OptimizerContext(ConnectContext context) {
+        this.connectContext = context;
         this.ruleSet = new RuleSet();
-        this.globalStateMgr = GlobalStateMgr.getCurrentState();
-        this.taskScheduler = TaskScheduler.create();
-        this.columnRefFactory = columnRefFactory;
-        this.sessionVariable = GlobalStateMgr.getCurrentState().getVariableMgr().newSessionVariable();
-        this.optimizerConfig = new OptimizerConfig();
-        this.queryId = UUID.randomUUID();
-        this.allLogicalOlapScanOperators = Collections.emptyList();
-    }
-
-    @VisibleForTesting
-    public OptimizerContext(Memo memo, ColumnRefFactory columnRefFactory, ConnectContext connectContext) {
-        this(memo, columnRefFactory, connectContext, OptimizerConfig.defaultConfig());
-    }
-
-    public OptimizerContext(Memo memo, ColumnRefFactory columnRefFactory, ConnectContext connectContext,
-                            OptimizerConfig optimizerConfig) {
-        this.memo = memo;
-        this.ruleSet = new RuleSet();
-        this.globalStateMgr = GlobalStateMgr.getCurrentState();
-        this.taskScheduler = TaskScheduler.create();
-        this.columnRefFactory = columnRefFactory;
-        this.queryId = connectContext.getQueryId();
-        this.sessionVariable = connectContext.getSessionVariable();
-        this.dumpInfo = connectContext.getDumpInfo();
-        this.currentSqlDbIds = connectContext.getCurrentSqlDbIds();
         this.cteContext = new CTEContext();
         cteContext.reset();
-        this.cteContext.setEnableCTE(sessionVariable.isCboCteReuse());
-        this.cteContext.setInlineCTERatio(sessionVariable.getCboCTERuseRatio());
-        this.cteContext.setMaxCTELimit(sessionVariable.getCboCTEMaxLimit());
-        this.optimizerConfig = optimizerConfig;
+        this.cteContext.setEnableCTE(getSessionVariable().isCboCteReuse());
+        this.cteContext.setInlineCTERatio(getSessionVariable().getCboCTERuseRatio());
+        this.cteContext.setMaxCTELimit(getSessionVariable().getCboCTEMaxLimit());
+
+        this.optimizerOptions = OptimizerOptions.defaultOpt();
     }
 
-    public Memo getMemo() {
-        return memo;
+    // ============================ Query ============================
+    public StatementBase getStatement() {
+        return statement;
     }
 
-    public RuleSet getRuleSet() {
-        return ruleSet;
+    public void setStatement(StatementBase statement) {
+        this.statement = statement;
     }
 
-    public GlobalStateMgr getCatalog() {
-        return globalStateMgr;
+    public ConnectContext getConnectContext() {
+        return connectContext;
     }
 
-    public TaskScheduler getTaskScheduler() {
-        return taskScheduler;
+    public void setConnectContext(ConnectContext connectContext) {
+        this.connectContext = connectContext;
+    }
+
+    public UUID getQueryId() {
+        return connectContext.getQueryId();
+    }
+
+    public final SessionVariable getSessionVariable() {
+        return connectContext.getSessionVariable();
+    }
+
+    public DumpInfo getDumpInfo() {
+        return connectContext.getDumpInfo();
+    }
+
+    public void setColumnRefFactory(ColumnRefFactory columnRefFactory) {
+        this.columnRefFactory = columnRefFactory;
     }
 
     public ColumnRefFactory getColumnRefFactory() {
         return columnRefFactory;
     }
 
-    public final SessionVariable getSessionVariable() {
-        return sessionVariable;
+    public Set<OlapTable> getQueryTables() {
+        return queryTables;
     }
 
-    public void setSessionVariable(SessionVariable sessionVariable) {
-        this.sessionVariable = sessionVariable;
+    public void setQueryTables(Set<OlapTable> queryTables) {
+        this.queryTables = queryTables;
     }
 
-    public DumpInfo getDumpInfo() {
-        return dumpInfo;
+    public void setUpdateTableId(long updateTableId) {
+        this.updateTableId = updateTableId;
     }
 
-    public Set<Long> getCurrentSqlDbIds() {
-        return currentSqlDbIds;
+    public long getUpdateTableId() {
+        return updateTableId;
+    }
+
+    public OptimizerOptions getOptimizerOptions() {
+        return optimizerOptions;
+    }
+
+    public void setOptimizerOptions(OptimizerOptions optimizerOptions) {
+        this.optimizerOptions = optimizerOptions;
+    }
+
+    // ============================ Optimizer ============================
+    public Memo getMemo() {
+        return memo;
+    }
+
+    public void setMemo(Memo memo) {
+        this.memo = memo;
+    }
+
+    public RuleSet getRuleSet() {
+        return ruleSet;
+    }
+
+    public TaskScheduler getTaskScheduler() {
+        return taskScheduler;
+    }
+
+    public void setTaskScheduler(TaskScheduler taskScheduler) {
+        this.taskScheduler = taskScheduler;
     }
 
     public CTEContext getCteContext() {
@@ -168,12 +185,37 @@ public class OptimizerContext {
         return currentTaskContext;
     }
 
-    public UUID getQueryId() {
-        return queryId;
+    public QueryMaterializationContext getQueryMaterializationContext() {
+        return queryMaterializationContext;
     }
 
-    public OptimizerConfig getOptimizerConfig() {
-        return optimizerConfig;
+    public MVTransformerContext getMvTransformerContext() {
+        return mvTransformerContext;
+    }
+
+    public void setMvTransformerContext(MVTransformerContext mvTransformerContext) {
+        this.mvTransformerContext = mvTransformerContext;
+    }
+
+    // ============================ Task Variables ============================
+    public JoinPredicatePushdown.JoinPredicatePushDownContext getJoinPushDownParams() {
+        return joinPredicatePushDownContext;
+    }
+
+    public boolean isObtainedFromInternalStatistics() {
+        return isObtainedFromInternalStatistics;
+    }
+
+    public void setObtainedFromInternalStatistics(boolean obtainedFromInternalStatistics) {
+        isObtainedFromInternalStatistics = obtainedFromInternalStatistics;
+    }
+
+    public void setInMemoPhase(boolean inMemoPhase) {
+        this.inMemoPhase = inMemoPhase;
+    }
+
+    public boolean isInMemoPhase() {
+        return this.inMemoPhase;
     }
 
     /**
@@ -185,22 +227,6 @@ public class OptimizerContext {
         return queryMaterializationContext.getValidCandidateMVs();
     }
 
-    public JoinPredicatePushdown.JoinPredicatePushDownContext getJoinPushDownParams() {
-        return joinPredicatePushDownContext;
-    }
-
-    public void setUpdateTableId(long updateTableId) {
-        this.updateTableId = updateTableId;
-    }
-
-    public long getUpdateTableId() {
-        return updateTableId;
-    }
-
-    public long optimizerElapsedMs() {
-        return optimizerTimer.elapsed(TimeUnit.MILLISECONDS);
-    }
-
     public Stopwatch getStopwatch(RuleType ruleType) {
         return ruleWatchMap.computeIfAbsent(ruleType, (k) -> Stopwatch.createStarted());
     }
@@ -208,84 +234,9 @@ public class OptimizerContext {
     public boolean ruleExhausted(RuleType ruleType) {
         Stopwatch watch = getStopwatch(ruleType);
         long elapsed = watch.elapsed(TimeUnit.MILLISECONDS);
-        long timeLimit = Math.min(sessionVariable.getOptimizerMaterializedViewTimeLimitMillis(),
-                sessionVariable.getOptimizerExecuteTimeout());
+        long timeLimit = Math.min(getSessionVariable().getOptimizerMaterializedViewTimeLimitMillis(),
+                getSessionVariable().getOptimizerExecuteTimeout());
         return elapsed > timeLimit;
-    }
-
-    public boolean isObtainedFromInternalStatistics() {
-        return isObtainedFromInternalStatistics;
-    }
-
-    public void setObtainedFromInternalStatistics(boolean obtainedFromInternalStatistics) {
-        isObtainedFromInternalStatistics = obtainedFromInternalStatistics;
-    }
-
-    /**
-     * Whether reach optimizer timeout
-     */
-    public boolean reachTimeout() {
-        long timeout = getSessionVariable().getOptimizerExecuteTimeout();
-        return optimizerElapsedMs() > timeout;
-    }
-
-    public Set<OlapTable> getQueryTables() {
-        return queryTables;
-    }
-
-    public void setQueryTables(Set<OlapTable> queryTables) {
-        this.queryTables = queryTables;
-    }
-
-    /**
-     * Throw exception if reach optimizer timeout
-     */
-    public void checkTimeout() {
-        if (!reachTimeout()) {
-            return;
-        }
-        Memo memo = getMemo();
-        Group group = memo == null ? null : memo.getRootGroup();
-        throw new StarRocksPlannerException("StarRocks planner use long time " + optimizerElapsedMs() +
-                " ms in " + (group == null ? "logical" : "memo") + " phase, This probably because " +
-                "1. FE Full GC, " +
-                "2. Hive external table fetch metadata took a long time, " +
-                "3. The SQL is very complex. " +
-                "You could " +
-                "1. adjust FE JVM config, " +
-                "2. try query again, " +
-                "3. enlarge new_planner_optimize_timeout session variable",
-                ErrorType.INTERNAL_ERROR);
-    }
-
-    public void setQueryMaterializationContext(QueryMaterializationContext queryMaterializationContext) {
-        this.queryMaterializationContext = queryMaterializationContext;
-    }
-
-    public QueryMaterializationContext getQueryMaterializationContext() {
-        return queryMaterializationContext;
-    }
-
-    public boolean isShortCircuit() {
-        return isShortCircuit;
-    }
-
-    public void setShortCircuit(boolean shortCircuit) {
-        isShortCircuit = shortCircuit;
-    }
-
-    public void clear() {
-        if (this.queryMaterializationContext != null) {
-            this.queryMaterializationContext.clear();
-        }
-    }
-
-    public void setInMemoPhase(boolean inMemoPhase) {
-        this.inMemoPhase = inMemoPhase;
-    }
-
-    public boolean isInMemoPhase() {
-        return this.inMemoPhase;
     }
 
     public List<IsNullPredicateOperator> getPushdownNotNullPredicates() {
@@ -305,11 +256,23 @@ public class OptimizerContext {
         return uniquePartitionIdGenerator++;
     }
 
-    public void setAllLogicalOlapScanOperators(List<LogicalOlapScanOperator> allScanOperators) {
-        this.allLogicalOlapScanOperators = allScanOperators;
-    }
-
-    public List<LogicalOlapScanOperator> getAllLogicalOlapScanOperators() {
-        return allLogicalOlapScanOperators;
+    /**
+     * Throw exception if reach optimizer timeout
+     */
+    public void checkTimeout() {
+        long timeout = getSessionVariable().getOptimizerExecuteTimeout();
+        long now = optimizerTimer.elapsed(TimeUnit.MILLISECONDS);
+        if (timeout > 0 && now > timeout) {
+            throw new StarRocksPlannerException("StarRocks planner use long time " + now +
+                    " ms in " + (inMemoPhase ? "memo" : "logical") + " phase, This probably because " +
+                    "1. FE Full GC, " +
+                    "2. Hive external table fetch metadata took a long time, " +
+                    "3. The SQL is very complex. " +
+                    "You could " +
+                    "1. adjust FE JVM config, " +
+                    "2. try query again, " +
+                    "3. enlarge new_planner_optimize_timeout session variable",
+                    ErrorType.INTERNAL_ERROR);
+        }
     }
 }
