+++ fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/transformation/materialization/MvRewriteStrategy.java
@@ -19,8 +19,8 @@ import com.starrocks.qe.ConnectContext;
 import com.starrocks.qe.SessionVariable;
 import com.starrocks.sql.optimizer.MaterializationContext;
 import com.starrocks.sql.optimizer.OptExpression;
-import com.starrocks.sql.optimizer.OptimizerConfig;
 import com.starrocks.sql.optimizer.OptimizerContext;
+import com.starrocks.sql.optimizer.OptimizerOptions;
 import com.starrocks.sql.optimizer.rule.RuleType;
 
 public class MvRewriteStrategy {
@@ -59,14 +59,14 @@ public class MvRewriteStrategy {
     public boolean enableMultiTableRewrite = false;
 
     static class MvStrategyArbitrator {
-        private final OptimizerConfig optimizerConfig;
+        private final OptimizerOptions optimizerOptions;
         private final OptimizerContext optimizerContext;
         private final SessionVariable sessionVariable;
 
         public MvStrategyArbitrator(OptimizerContext optimizerContext,
                                     ConnectContext connectContext) {
             this.optimizerContext = optimizerContext;
-            this.optimizerConfig = optimizerContext.getOptimizerConfig();
+            this.optimizerOptions = optimizerContext.getOptimizerOptions();
             // from connectContext rather than optimizerContext
             this.sessionVariable = connectContext.getSessionVariable();
         }
@@ -84,8 +84,8 @@ public class MvRewriteStrategy {
                     optimizerContext.getCandidateMvs().isEmpty()) {
                 return false;
             }
-            if (optimizerConfig.isRuleDisable(RuleType.GP_SINGLE_TABLE_MV_REWRITE) &&
-                    optimizerConfig.isRuleDisable(RuleType.GP_MULTI_TABLE_MV_REWRITE)) {
+            if (optimizerOptions.isRuleDisable(RuleType.GP_SINGLE_TABLE_MV_REWRITE) &&
+                    optimizerOptions.isRuleDisable(RuleType.GP_MULTI_TABLE_MV_REWRITE)) {
                 return false;
             }
             return true;
@@ -100,7 +100,7 @@ public class MvRewriteStrategy {
 
         private boolean isEnableRBOSingleTableRewrite(OptExpression queryPlan) {
             // if disable single mv rewrite, return false.
-            if (optimizerConfig.isRuleDisable(RuleType.GP_SINGLE_TABLE_MV_REWRITE)) {
+            if (optimizerOptions.isRuleDisable(RuleType.GP_SINGLE_TABLE_MV_REWRITE)) {
                 return false;
             }
             // If query only has one table use single table rewrite, view delta only rewrites multi-tables query.
