+++ fe/fe-core/src/main/java/com/starrocks/sql/optimizer/task/TaskScheduler.java
@@ -16,10 +16,6 @@ package com.starrocks.sql.optimizer.task;
 
 import com.starrocks.common.profile.Timer;
 import com.starrocks.common.profile.Tracers;
-import com.starrocks.sql.common.ErrorType;
-import com.starrocks.sql.common.StarRocksPlannerException;
-import com.starrocks.sql.optimizer.Group;
-import com.starrocks.sql.optimizer.Memo;
 import com.starrocks.sql.optimizer.OptExpression;
 import com.starrocks.sql.optimizer.rule.Rule;
 
@@ -39,29 +35,8 @@ public class TaskScheduler {
     }
 
     public void executeTasks(TaskContext context) {
-        long timeout = context.getOptimizerContext().getSessionVariable().getOptimizerExecuteTimeout();
         while (!tasks.empty()) {
-            long watch = context.getOptimizerContext().optimizerElapsedMs();
-            if (timeout > 0 && watch > timeout) {
-                // Should have at least one valid plan
-                // group will be null when in rewrite phase
-                // memo may be null for rule-based optimizer
-                Memo memo = context.getOptimizerContext().getMemo();
-                Group group = memo == null ? null : memo.getRootGroup();
-                if (group == null || !group.hasBestExpression(context.getRequiredProperty())) {
-                    throw new StarRocksPlannerException("StarRocks planner use long time " + timeout +
-                            " ms in " + (group == null ? "logical" : "memo") + " phase, This probably because " +
-                            "1. FE Full GC, " +
-                            "2. Hive external table fetch metadata took a long time, " +
-                            "3. The SQL is very complex. " +
-                            "You could " +
-                            "1. adjust FE JVM config, " +
-                            "2. try query again, " +
-                            "3. enlarge new_planner_optimize_timeout session variable",
-                            ErrorType.INTERNAL_ERROR);
-                }
-                break;
-            }
+            context.getOptimizerContext().checkTimeout();
             OptimizerTask task = tasks.pop();
             context.getOptimizerContext().setTaskContext(context);
             try (Timer ignore = Tracers.watchScope(Tracers.Module.OPTIMIZER, task.getClass().getSimpleName())) {
