+++ fe/fe-core/src/main/java/com/starrocks/sql/plan/ExecPlan.java
@@ -67,6 +67,8 @@ public class ExecPlan {
 
     private List<Integer> collectExecStatsIds;
 
+    private final boolean isShortCircuit;
+
     @VisibleForTesting
     public ExecPlan() {
         connectContext = new ConnectContext();
@@ -74,14 +76,16 @@ public class ExecPlan {
         colNames = new ArrayList<>();
         physicalPlan = null;
         outputColumns = new ArrayList<>();
+        isShortCircuit = false;
     }
 
     public ExecPlan(ConnectContext connectContext, List<String> colNames,
-                    OptExpression physicalPlan, List<ColumnRefOperator> outputColumns) {
+                    OptExpression physicalPlan, List<ColumnRefOperator> outputColumns, boolean isShortCircuit) {
         this.connectContext = connectContext;
         this.colNames = colNames;
         this.physicalPlan = physicalPlan;
         this.outputColumns = outputColumns;
+        this.isShortCircuit = isShortCircuit;
     }
 
     // for broker load plan
@@ -91,6 +95,7 @@ public class ExecPlan {
         this.physicalPlan = null;
         this.outputColumns = new ArrayList<>();
         this.fragments.addAll(fragments);
+        this.isShortCircuit = false;
     }
 
     public ConnectContext getConnectContext() {
@@ -270,5 +275,9 @@ public class ExecPlan {
 
     public void setCollectExecStatsIds(List<Integer> collectExecStatsIds) {
         this.collectExecStatsIds = collectExecStatsIds;
+    }
+
+    public boolean isShortCircuit() {
+        return isShortCircuit;
     }
 }
