+++ fe/fe-core/src/test/java/com/starrocks/sql/optimizer/OptimizerTaskTest.java
@@ -39,7 +39,6 @@ import com.starrocks.sql.optimizer.base.ColumnRefSet;
 import com.starrocks.sql.optimizer.base.DistributionSpec;
 import com.starrocks.sql.optimizer.base.HashDistributionDesc;
 import com.starrocks.sql.optimizer.base.Ordering;
-import com.starrocks.sql.optimizer.base.PhysicalPropertySet;
 import com.starrocks.sql.optimizer.dump.MockDumpInfo;
 import com.starrocks.sql.optimizer.operator.AggType;
 import com.starrocks.sql.optimizer.operator.Operator;
@@ -169,9 +168,8 @@ public class OptimizerTaskTest {
             }
         };
 
-        Optimizer optimizer = new Optimizer();
-        optimizer.optimize(ctx, logicOperatorTree, new PhysicalPropertySet(), new ColumnRefSet(),
-                columnRefFactory);
+        Optimizer optimizer = OptimizerFactory.create(OptimizerFactory.mockContext(ctx, columnRefFactory));
+        optimizer.optimize(logicOperatorTree, new ColumnRefSet());
         Memo memo = optimizer.getContext().getMemo();
         assertEquals(3, memo.getGroups().size());
         assertEquals(8, memo.getGroupExpressions().size());
@@ -257,9 +255,8 @@ public class OptimizerTaskTest {
             }
         };
 
-        Optimizer optimizer = new Optimizer();
-        optimizer.optimize(ctx, topJoin, new PhysicalPropertySet(), new ColumnRefSet(),
-                columnRefFactory);
+        Optimizer optimizer = OptimizerFactory.create(OptimizerFactory.mockContext(ctx, columnRefFactory));
+        optimizer.optimize(topJoin, new ColumnRefSet());
 
         Memo memo = optimizer.getContext().getMemo();
 
@@ -379,9 +376,8 @@ public class OptimizerTaskTest {
                                 scan4ColumnMap, Maps.newHashMap(), null,
                                 -1, null)));
 
-        Optimizer optimizer = new Optimizer();
-        optimizer.optimize(ctx, topJoin, new PhysicalPropertySet(), new ColumnRefSet(),
-                columnRefFactory);
+        Optimizer optimizer = OptimizerFactory.create(OptimizerFactory.mockContext(ctx, columnRefFactory));
+        optimizer.optimize(topJoin, new ColumnRefSet());
     }
 
     @Test
@@ -475,9 +471,8 @@ public class OptimizerTaskTest {
             }
         };
 
-        Optimizer optimizer = new Optimizer();
-        optimizer.optimize(ctx, topJoin, new PhysicalPropertySet(), new ColumnRefSet(),
-                columnRefFactory);
+        Optimizer optimizer = OptimizerFactory.create(OptimizerFactory.mockContext(ctx, columnRefFactory));
+        optimizer.optimize(topJoin, new ColumnRefSet());
     }
 
     @Test
@@ -622,9 +617,8 @@ public class OptimizerTaskTest {
             }
         };
 
-        Optimizer optimizer = new Optimizer();
-        optimizer.optimize(ctx, topJoin, new PhysicalPropertySet(), new ColumnRefSet(),
-                columnRefFactory);
+        Optimizer optimizer = OptimizerFactory.create(OptimizerFactory.mockContext(ctx, columnRefFactory));
+        optimizer.optimize(topJoin, new ColumnRefSet());
     }
 
     @Test
@@ -670,8 +664,9 @@ public class OptimizerTaskTest {
         outputColumns.addAll(outputColumns2);
         ColumnRefSet outputColumnsSet = new ColumnRefSet(outputColumns);
 
-        Optimizer optimizer = new Optimizer();
-        optimizer.optimize(ctx, logicOperatorTree, new PhysicalPropertySet(), outputColumnsSet, columnRefFactory);
+        Optimizer optimizer = OptimizerFactory.create(OptimizerFactory.mockContext(ctx, columnRefFactory));
+        optimizer.optimize(logicOperatorTree, outputColumnsSet);
+
         Memo memo = optimizer.getContext().getMemo();
 
         MemoStatusChecker checker = new MemoStatusChecker(memo, 2, outputColumnsSet);
@@ -715,10 +710,10 @@ public class OptimizerTaskTest {
                         new LogicalOlapScanOperator(olapTable3, scanColumnMap,
                                 Maps.newHashMap(), null, -1, null)));
 
-        Optimizer optimizer = new Optimizer();
-        OptExpression physicalTree = optimizer.optimize(ctx, topJoin, new PhysicalPropertySet(),
-                new ColumnRefSet(Lists.newArrayList(column1)),
-                columnRefFactory);
+        Optimizer optimizer = OptimizerFactory.create(OptimizerFactory.mockContext(ctx, columnRefFactory));
+        OptExpression physicalTree = optimizer.optimize(topJoin,
+                new ColumnRefSet(List.of(column1)));
+
         assertEquals(physicalTree.getOp().getOpType(), OperatorType.PHYSICAL_NESTLOOP_JOIN);
         assertEquals(physicalTree.inputAt(0).getOp().getOpType(), OperatorType.PHYSICAL_NESTLOOP_JOIN);
         assertEquals(physicalTree.inputAt(1).getOp().getOpType(), OperatorType.PHYSICAL_DISTRIBUTION);
@@ -750,9 +745,9 @@ public class OptimizerTaskTest {
             }
         };
 
-        Optimizer optimizer = new Optimizer();
-        OptExpression physicalTree = optimizer.optimize(ctx, expression, new PhysicalPropertySet(), new ColumnRefSet(),
-                columnRefFactory);
+        Optimizer optimizer = OptimizerFactory.create(OptimizerFactory.mockContext(ctx, columnRefFactory));
+        OptExpression physicalTree = optimizer.optimize(expression, new ColumnRefSet());
+
         Operator root = physicalTree.getOp();
         assertEquals(root.getOpType(), OperatorType.PHYSICAL_LIMIT);
     }
@@ -783,9 +778,9 @@ public class OptimizerTaskTest {
                         new LogicalOlapScanOperator(olapTable1, scanColumnMap, Maps.newHashMap(), null,
                                 -1, null)));
 
-        Optimizer optimizer = new Optimizer();
-        OptExpression physicalTree = optimizer.optimize(ctx, expression, new PhysicalPropertySet(),
-                new ColumnRefSet(outputColumns1), columnRefFactory);
+        Optimizer optimizer = OptimizerFactory.create(OptimizerFactory.mockContext(ctx, columnRefFactory));
+        OptExpression physicalTree = optimizer.optimize(expression,
+                new ColumnRefSet(outputColumns1));
 
         assertEquals(OperatorType.PHYSICAL_OLAP_SCAN, physicalTree.getOp().getOpType());
         PhysicalOlapScanOperator physicalOlapScan = (PhysicalOlapScanOperator) physicalTree.getOp();
@@ -829,9 +824,10 @@ public class OptimizerTaskTest {
             }
         };
 
-        Optimizer optimizer = new Optimizer();
-        OptExpression physicalTree = optimizer.optimize(ctx, expression, new PhysicalPropertySet(),
-                new ColumnRefSet(outputColumns1), columnRefFactory);
+        Optimizer optimizer = OptimizerFactory.create(
+                OptimizerFactory.mockContext(ctx, columnRefFactory));
+        OptExpression physicalTree = optimizer.optimize(expression,
+                new ColumnRefSet(outputColumns1));
 
         assertNotNull(physicalTree.getOp().getProjection());
         assertEquals(physicalTree.getOp().getOpType(), OperatorType.PHYSICAL_OLAP_SCAN);
@@ -901,9 +897,9 @@ public class OptimizerTaskTest {
 
         ColumnRefSet outputColumns = new ColumnRefSet(column2.getId());
 
-        Optimizer optimizer = new Optimizer();
         try {
-            optimizer.optimize(ctx, expression, new PhysicalPropertySet(), outputColumns, columnRefFactory);
+            Optimizer optimizer = OptimizerFactory.create(OptimizerFactory.mockContext(ctx, columnRefFactory));
+            optimizer.optimize(expression, outputColumns);
             fail("invalid plan. can not optimize success");
         } catch (Exception e) {
             Assert.assertTrue(e.getMessage(), e.getMessage().contains("Type check failed. the type of arg 2: t2 in " +
@@ -917,8 +913,10 @@ public class OptimizerTaskTest {
             tmp.getSessionVariable().setJoinImplementationMode("auto");
             tmp.getSessionVariable().setEnablePlanValidation(false);
             tmp.setDumpInfo(new MockDumpInfo());
-            OptExpression expression1 = optimizer.optimize(tmp, expression, new PhysicalPropertySet(), outputColumns,
-                    columnRefFactory);
+
+            Optimizer optimizer = OptimizerFactory.create(
+                    OptimizerFactory.mockContext(ctx, columnRefFactory));
+            OptExpression expression1 = optimizer.optimize(expression, outputColumns);
             Map<ColumnRefOperator, CallOperator> aggs =
                     ((PhysicalHashAggregateOperator) expression1.getOp()).getAggregations();
             assertEquals(Type.INT, column2.getType());
@@ -994,9 +992,10 @@ public class OptimizerTaskTest {
 
         ColumnRefSet outputColumns = new ColumnRefSet(column5.getId());
 
-        Optimizer optimizer = new Optimizer();
         try {
-            optimizer.optimize(ctx, expression, new PhysicalPropertySet(), outputColumns, columnRefFactory);
+            Optimizer optimizer = OptimizerFactory.create(OptimizerFactory.mockContext(ctx, columnRefFactory));
+            optimizer.optimize(expression, outputColumns);
+
             fail("invalid plan. can not optimize success");
         } catch (Exception e) {
             Assert.assertTrue(e.getMessage(), e.getMessage().contains("Type check failed. the type of arg 5: t5 in " +
@@ -1047,9 +1046,11 @@ public class OptimizerTaskTest {
             }
         };
 
-        Optimizer optimizer = new Optimizer();
         try {
-            optimizer.optimize(ctx, expression, new PhysicalPropertySet(), outputColumns, columnRefFactory);
+            Optimizer optimizer = OptimizerFactory.create(
+                    OptimizerFactory.mockContext(ctx, columnRefFactory));
+            optimizer.optimize(expression, outputColumns);
+
             fail("invalid plan. can not optimize success");
         } catch (Exception e) {
             Assert.assertTrue(e.getMessage(), e.getMessage().contains("Type check failed. the type of arg 3: t3 in " +
@@ -1104,9 +1105,11 @@ public class OptimizerTaskTest {
             }
         };
 
-        Optimizer optimizer = new Optimizer();
         try {
-            optimizer.optimize(ctx, expression, new PhysicalPropertySet(), outputColumns, columnRefFactory);
+            Optimizer optimizer = OptimizerFactory.create(
+                    OptimizerFactory.mockContext(ctx, columnRefFactory));
+            optimizer.optimize(expression, outputColumns);
+
             fail("invalid plan. can not optimize success");
         } catch (Exception e) {
             Assert.assertTrue(e.getMessage(), e.getMessage().contains("Type check failed. the type of arg 3: t3 in " +
@@ -1154,9 +1157,9 @@ public class OptimizerTaskTest {
 
         ColumnRefSet outputColumns = new ColumnRefSet(column2.getId());
 
-        Optimizer optimizer = new Optimizer();
-        OptExpression physicalTree = optimizer.optimize(ctx, expression, new PhysicalPropertySet(),
-                outputColumns, columnRefFactory);
+        Optimizer optimizer = OptimizerFactory.create(
+                OptimizerFactory.mockContext(ctx, columnRefFactory));
+        OptExpression physicalTree = optimizer.optimize(expression, outputColumns);
 
         assertEquals(physicalTree.getLogicalProperty().getOutputColumns(), new ColumnRefSet(column2.getId()));
 
@@ -1209,9 +1212,9 @@ public class OptimizerTaskTest {
 
         ColumnRefSet outputColumns = new ColumnRefSet(column1.getId());
 
-        Optimizer optimizer = new Optimizer();
-        OptExpression physicalTree = optimizer.optimize(ctx, expression, new PhysicalPropertySet(),
-                outputColumns, columnRefFactory);
+        Optimizer optimizer = OptimizerFactory.create(
+                OptimizerFactory.mockContext(ctx, columnRefFactory));
+        OptExpression physicalTree = optimizer.optimize(expression, outputColumns);
 
         Operator operator = physicalTree.getOp();
         assertEquals(operator.getOpType(), OperatorType.PHYSICAL_HASH_AGG);
@@ -1281,9 +1284,11 @@ public class OptimizerTaskTest {
             }
         };
 
-        Optimizer optimizer = new Optimizer();
         try {
-            optimizer.optimize(ctx, expression, new PhysicalPropertySet(), outputColumns, columnRefFactory);
+            Optimizer optimizer = OptimizerFactory.create(
+                    OptimizerFactory.mockContext(ctx, columnRefFactory));
+            optimizer.optimize(expression, outputColumns);
+
             fail("invalid plan. can not optimize success");
         } catch (Exception e) {
             Assert.assertTrue(e.getMessage(), e.getMessage().contains("Type check failed. the type of arg 3: t3 in " +
@@ -1342,9 +1347,10 @@ public class OptimizerTaskTest {
 
         ColumnRefSet outputColumns = new ColumnRefSet(Lists.newArrayList(column3, column2));
 
-        Optimizer optimizer = new Optimizer();
         try {
-            optimizer.optimize(ctx, expression, new PhysicalPropertySet(), outputColumns, columnRefFactory);
+            Optimizer optimizer = OptimizerFactory.create(
+                    OptimizerFactory.mockContext(ctx, columnRefFactory));
+            optimizer.optimize(expression, outputColumns);
             fail("invalid plan. can not optimize success");
         } catch (Exception e) {
             Assert.assertTrue(e.getMessage(), e.getMessage().contains("Type check failed. the type of arg 3: t3 in " +
@@ -1403,9 +1409,10 @@ public class OptimizerTaskTest {
 
         ColumnRefSet outputColumns = new ColumnRefSet(Lists.newArrayList(column3));
 
-        Optimizer optimizer = new Optimizer();
         try {
-            optimizer.optimize(ctx, expression, new PhysicalPropertySet(), outputColumns, columnRefFactory);
+            Optimizer optimizer = OptimizerFactory.create(OptimizerFactory.mockContext(ctx, columnRefFactory));
+            optimizer.optimize(expression, outputColumns);
+
             fail("invalid plan. can not optimize success");
         } catch (Exception e) {
             Assert.assertTrue(e.getMessage(), e.getMessage().contains("Type check failed. the type of arg 3: t3 in " +
@@ -1458,9 +1465,8 @@ public class OptimizerTaskTest {
 
         ColumnRefSet outputColumns = new ColumnRefSet(column2.getId());
 
-        Optimizer optimizer = new Optimizer();
-        OptExpression physicalTree = optimizer.optimize(ctx, topProject, new PhysicalPropertySet(),
-                outputColumns, columnRefFactory);
+        Optimizer optimizer = OptimizerFactory.create(OptimizerFactory.mockContext(ctx, columnRefFactory));
+        OptExpression physicalTree = optimizer.optimize(topProject, outputColumns);
 
         Operator operator = physicalTree.getOp();
         assertEquals(operator.getOpType(), OperatorType.PHYSICAL_HASH_AGG);
@@ -1524,9 +1530,10 @@ public class OptimizerTaskTest {
 
         ColumnRefSet outputColumns = new ColumnRefSet(Lists.newArrayList(column6));
 
-        Optimizer optimizer = new Optimizer();
         try {
-            optimizer.optimize(ctx, root, new PhysicalPropertySet(), outputColumns, columnRefFactory);
+            Optimizer optimizer = OptimizerFactory.create(OptimizerFactory.mockContext(ctx, columnRefFactory));
+            optimizer.optimize(root, outputColumns);
+
             fail("invalid plan. can not optimize success");
         } catch (Exception e) {
             Assert.assertTrue(e.getMessage(), e.getMessage().contains("Type check failed. the type of arg 5: t5 in " +
@@ -1583,9 +1590,9 @@ public class OptimizerTaskTest {
 
         ColumnRefSet outputColumns = new ColumnRefSet(Lists.newArrayList(column4));
 
-        Optimizer optimizer = new Optimizer();
-        OptExpression physicalTree = optimizer.optimize(ctx, root, new PhysicalPropertySet(),
-                outputColumns, columnRefFactory);
+        Optimizer optimizer = OptimizerFactory.create(
+                OptimizerFactory.mockContext(ctx, columnRefFactory));
+        OptExpression physicalTree = optimizer.optimize(root, outputColumns);
 
         Operator operator = physicalTree.getOp();
         assertEquals(operator.getOpType(), OperatorType.PHYSICAL_HASH_AGG);
@@ -1636,9 +1643,8 @@ public class OptimizerTaskTest {
                                         null, -1,
                                         null))));
 
-        Optimizer optimizer = new Optimizer();
-        OptExpression physicalTree = optimizer.optimize(ctx, expression, new PhysicalPropertySet(),
-                new ColumnRefSet(outputColumns), columnRefFactory);
+        Optimizer optimizer = OptimizerFactory.create(OptimizerFactory.mockContext(ctx, columnRefFactory));
+        OptExpression physicalTree = optimizer.optimize(expression, new ColumnRefSet(outputColumns));
 
         assertNotNull(physicalTree.getOp().getProjection());
         assertEquals(physicalTree.getOp().getOpType(), OperatorType.PHYSICAL_OLAP_SCAN);
@@ -1685,9 +1691,8 @@ public class OptimizerTaskTest {
 
         OptExpression filter = OptExpression.create(new LogicalFilterOperator(predicate), project2);
 
-        Optimizer optimizer = new Optimizer();
-        OptExpression physicalTree = optimizer.optimize(ctx, filter, new PhysicalPropertySet(),
-                new ColumnRefSet(outputColumns), columnRefFactory);
+        Optimizer optimizer = OptimizerFactory.create(OptimizerFactory.mockContext(ctx, columnRefFactory));
+        OptExpression physicalTree = optimizer.optimize(filter, new ColumnRefSet(outputColumns));
 
         assertNotNull(physicalTree.getOp().getProjection());
         Projection pp = physicalTree.getOp().getProjection();
@@ -1748,9 +1753,9 @@ public class OptimizerTaskTest {
                         new LogicalOlapScanOperator(olapTable1, scanColumnMap, Maps.newHashMap(), null, -1,
                                 null)));
 
-        Optimizer optimizer = new Optimizer();
-        OptExpression physicalTree = optimizer.optimize(ctx, expression, new PhysicalPropertySet(),
-                new ColumnRefSet(outputColumns), columnRefFactory);
+        Optimizer optimizer = OptimizerFactory.create(
+                OptimizerFactory.mockContext(ctx, columnRefFactory));
+        OptExpression physicalTree = optimizer.optimize(expression, new ColumnRefSet(outputColumns));
 
         PhysicalOlapScanOperator olapScanOperator = (PhysicalOlapScanOperator) physicalTree.getOp();
         Projection projection = olapScanOperator.getProjection();
@@ -1834,9 +1839,8 @@ public class OptimizerTaskTest {
                 OptExpression.create(scan1),
                 OptExpression.create(scan2));
 
-        Optimizer optimizer = new Optimizer();
-        OptExpression physicalTree = optimizer.optimize(ctx, expression, new PhysicalPropertySet(),
-                new ColumnRefSet(outputColumns), columnRefFactory);
+        Optimizer optimizer = OptimizerFactory.create(OptimizerFactory.mockContext(ctx, columnRefFactory));
+        OptExpression physicalTree = optimizer.optimize(expression, new ColumnRefSet(outputColumns));
     }
 
     @Test
@@ -1932,9 +1936,8 @@ public class OptimizerTaskTest {
                 OptExpression.create(scan2));
         OptExpression topJoin = OptExpression.create(join2,
                 join1, OptExpression.create(scan3));
-        Optimizer optimizer = new Optimizer();
-        OptExpression physicalTree = optimizer.optimize(ctx, topJoin, new PhysicalPropertySet(),
-                new ColumnRefSet(outputColumns), columnRefFactory);
+        Optimizer optimizer = OptimizerFactory.create(OptimizerFactory.mockContext(ctx, columnRefFactory));
+        OptExpression physicalTree = optimizer.optimize(topJoin, new ColumnRefSet(outputColumns));
     }
 
     @Test
@@ -2031,9 +2034,10 @@ public class OptimizerTaskTest {
                 OptExpression.create(scan1),
                 OptExpression.create(scan2));
 
-        Optimizer optimizer = new Optimizer();
-        OptExpression physicalTree = optimizer.optimize(ctx, expression, new PhysicalPropertySet(),
-                new ColumnRefSet(outputColumns), columnRefFactory);
+        Optimizer optimizer = OptimizerFactory.create(
+                OptimizerFactory.mockContext(ctx, columnRefFactory));
+        OptExpression physicalTree = optimizer.optimize(expression, new ColumnRefSet(outputColumns));
+
         assertEquals(physicalTree.getInputs().get(1).getOp().getOpType(), OperatorType.PHYSICAL_DISTRIBUTION);
         PhysicalDistributionOperator rightOperator =
                 (PhysicalDistributionOperator) physicalTree.getInputs().get(1).getOp();
@@ -2152,9 +2156,10 @@ public class OptimizerTaskTest {
                 OptExpression.create(scan1),
                 OptExpression.create(scan2));
 
-        Optimizer optimizer = new Optimizer();
-        OptExpression physicalTree = optimizer.optimize(ctx, expression, new PhysicalPropertySet(),
-                new ColumnRefSet(outputColumns), columnRefFactory);
+        Optimizer optimizer = OptimizerFactory.create(
+                OptimizerFactory.mockContext(ctx, columnRefFactory));
+        OptExpression physicalTree = optimizer.optimize(expression, new ColumnRefSet(outputColumns));
+
         assertEquals(physicalTree.getInputs().get(1).getOp().getOpType(), OperatorType.PHYSICAL_DISTRIBUTION);
         PhysicalDistributionOperator rightOperator =
                 (PhysicalDistributionOperator) physicalTree.getInputs().get(1).getOp();
@@ -2277,9 +2282,9 @@ public class OptimizerTaskTest {
                 OptExpression.create(scan1),
                 OptExpression.create(scan2));
 
-        Optimizer optimizer = new Optimizer();
-        OptExpression physicalTree = optimizer.optimize(ctx, expression, new PhysicalPropertySet(),
-                new ColumnRefSet(outputColumns), columnRefFactory);
+        Optimizer optimizer = OptimizerFactory.create(
+                OptimizerFactory.mockContext(ctx, columnRefFactory));
+        OptExpression physicalTree = optimizer.optimize(expression, new ColumnRefSet(outputColumns));
         assertEquals(physicalTree.getInputs().get(1).getOp().getOpType(), OperatorType.PHYSICAL_DISTRIBUTION);
         PhysicalDistributionOperator rightOperator =
                 (PhysicalDistributionOperator) physicalTree.getInputs().get(1).getOp();
@@ -2334,7 +2339,6 @@ public class OptimizerTaskTest {
             }
         };
 
-        optimizer = new Optimizer();
         expression = OptExpression.create(new LogicalJoinOperator(JoinOperator.INNER_JOIN, predicate),
                 OptExpression.create(new LogicalOlapScanOperator(olapTable1, scan1ColumnMap, Maps.newHashMap(),
                         DistributionSpec.createHashDistributionSpec(
@@ -2344,8 +2348,10 @@ public class OptimizerTaskTest {
                         DistributionSpec.createHashDistributionSpec(
                                 new HashDistributionDesc(Lists.newArrayList(this.column4.getId()),
                                         HashDistributionDesc.SourceType.LOCAL)), -1, null)));
-        physicalTree = optimizer.optimize(ctx, expression, new PhysicalPropertySet(),
-                new ColumnRefSet(outputColumns), columnRefFactory);
+        optimizer = OptimizerFactory.create(
+                OptimizerFactory.mockContext(ctx, columnRefFactory));
+        physicalTree = optimizer.optimize(expression, new ColumnRefSet(outputColumns));
+
         assertEquals(physicalTree.getInputs().get(1).getOp().getOpType(), OperatorType.PHYSICAL_DISTRIBUTION);
         rightOperator = (PhysicalDistributionOperator) physicalTree.getInputs().get(1).getOp();
         assertEquals(rightOperator.getDistributionSpec().getType(), DistributionSpec.DistributionType.BROADCAST);
@@ -2395,7 +2401,7 @@ public class OptimizerTaskTest {
 
         SplitTwoPhaseAggRule splitTwoPhaseAggRule = SplitTwoPhaseAggRule.getInstance();
         List<OptExpression> list = splitTwoPhaseAggRule.transform(
-                expression, new OptimizerContext(new Memo(), new ColumnRefFactory()));
+                expression, OptimizerFactory.mockContext(new ColumnRefFactory()));
 
         assertEquals(OperatorType.LOGICAL_AGGR, list.get(0).getOp().getOpType());
         LogicalAggregationOperator result = (LogicalAggregationOperator) list.get(0).getOp();
