+++ fe/fe-core/src/test/java/com/starrocks/sql/optimizer/rule/implementation/OlapScanImplementationRuleTest.java
@@ -17,9 +17,8 @@ package com.starrocks.sql.optimizer.rule.implementation;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import com.starrocks.catalog.OlapTable;
-import com.starrocks.sql.optimizer.Memo;
 import com.starrocks.sql.optimizer.OptExpression;
-import com.starrocks.sql.optimizer.OptimizerContext;
+import com.starrocks.sql.optimizer.OptimizerFactory;
 import com.starrocks.sql.optimizer.base.ColumnRefFactory;
 import com.starrocks.sql.optimizer.operator.logical.LogicalOlapScanOperator;
 import com.starrocks.sql.optimizer.operator.physical.PhysicalOlapScanOperator;
@@ -41,8 +40,8 @@ public class OlapScanImplementationRuleTest {
                 false, Lists.newArrayList(4L), null, null, false);
 
         List<OptExpression> output =
-                new OlapScanImplementationRule().transform(new OptExpression(logical), new OptimizerContext(
-                        new Memo(), new ColumnRefFactory()));
+                new OlapScanImplementationRule().transform(new OptExpression(logical),
+                        OptimizerFactory.mockContext(new ColumnRefFactory()));
 
         assertEquals(1, output.size());
